refactor: extract delivery log tests to stay under 1000-line limit
