+++ crates/rs-core/src/db/mod.rs
@@ -18,6 +18,9 @@ mod tests;
 #[cfg(test)]
 mod template_tests;
 
+#[cfg(test)]
+mod delivery_log_tests;
+
 /// Create a SQLite connection pool.
 pub async fn create_pool(db_path: &Path) -> Result<SqlitePool> {
     let url = format!("sqlite:{}?mode=rwc", db_path.display());
